+++ src/component/cpu/mc68020.hpp
@@ -24,6 +24,13 @@ namespace Motion
             this->moiraCpu = bridge;
         }
 
+        // for the coherent stack window
+
+        uint32_t GetStack32(uint32_t offset) override
+        {
+            return AddrSpace::ReadU32(moiraCpu->reg.sp + (offset << 2));
+        };
+
     private: 
         char disasmBuf[MOIRA_DISASM_BUF_SIZE] = {0};
         MC68020MoiraBridge* moiraCpu;