+++ src/component/multibus/multibus.hpp
@@ -1,0 +1,12 @@
+/* 
+    m  o  t  i  o  n
+    The SGI Emulator
+
+    Copyright (c)2026 starfrost
+
+    multibus.hpp: The Intel Multibus bus.
+    This is the bus that is used for everything on the early IRISes, and most things on the later IRIS 3000.
+
+    Most likely it was inheited from the SUN design that SGI bought back in '82. PM1/PM2/IP1 boards are fully multibus, IP2 boards,
+    except for the CPU, GPU and FPU (?) (thees use their own private bus.)
+*/